Seventh Commit - commentController: comment on a post, get all comments of a post and delete a comment.

+++ migrations/20220910084917-create-post.js
@@ -36,7 +36,7 @@ module.exports = {
         defaultValue: 0,
       },
       rating: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT(5, 4),
         allowNull: false,
         defaultValue: 1.2,
       },
